+++ include/packet.h
@@ -17,6 +17,6 @@ struct Client;
  */
 
 extern int server_dopacket(struct Client* cptr, const char* buffer, int length);
-extern int client_dopacket(struct Client* cptr, size_t length);
+extern int client_dopacket(struct Client* cptr, unsigned int length);
 
 #endif /* INCLUDED_packet_h */
